+++ src/XRP/wallet-factory.ts
@@ -1,3 +1,5 @@
+import * as rippleKeyPair from 'ripple-keypairs'
+
 import Utils from '../Common/utils'
 
 import Wallet from './wallet'
@@ -21,6 +23,21 @@ export default class WalletFactory {
   public constructor(network: XrplNetwork) {
     this.network = network
     this.isTest = XrpUtils.isTestNetwork(network)
+  }
+
+  /**
+   * Generate a new wallet from the given seed.
+   *
+   * @param seed - The given seed for the wallet.
+   * @returns A new wallet from the given seed, or undefined if the seed was invalid.
+   */
+  public walletFromSeed(seed: string): Wallet | undefined {
+    try {
+      const keyPair = rippleKeyPair.deriveKeypair(seed)
+      return new Wallet(keyPair.publicKey, keyPair.privateKey, this.isTest)
+    } catch {
+      return undefined
+    }
   }
 
   /**