+++ CHANGELOG.md
@@ -7,12 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Added
+### Deprecated
 
+- XRP specific functionality on the `Utils` class is deprecated. Use the new `XrpUtils` class instead.
 - A new method, `walletFromSeed` in `WalletFactory` encapsulates functionality for creating a `Wallet` object from a seed.
 
 ### Deprecated
 
+- XRP-specific functionality on the `Utils` class is deprecated. Use the new `XrpUtils` class instead.
 - The `walletFromSeed` method of the `Wallet` class is deprecated. Please use the new `WalletFactory` to generate `Wallet`s from seeds.
 
 ## [5.1.0] - 2020-06-01