+++ CHANGELOG.md
@@ -11,11 +11,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - XRP specific functionality on the `Utils` class is deprecated. Use the new `XrpUtils` class instead.
 - A new method, `walletFromSeed` in `WalletFactory` encapsulates functionality for creating a `Wallet` object from a seed.
+- A new enum, `XrplNetwork`, identifies the XRP Ledger network that components are connected to.
+- A new method, `isTestNetwork`, is provided by `XrpUtils` to identify whether a given `XrplNetwork` is a test network.
 
 ### Deprecated
 
 - XRP-specific functionality on the `Utils` class is deprecated. Use the new `XrpUtils` class instead.
 - The `walletFromSeed` method of the `Wallet` class is deprecated. Please use the new `WalletFactory` to generate `Wallet`s from seeds.
+
+### Breaking Changes
+
+- `WalletFactory`'s constructor takes an `XrplNetwork` enum instead of a boolean indicating if the network is a test network. Clients should pass `XrplNetwork.Test` if they were previously passing `true`, otherwise they should pass `XrplNetwork.Main`.
 
 ## [5.1.0] - 2020-06-01